src/requirements.txt
django
psycopg2-binary
requests_cache
djangorestframework
django-filter
markdown